+++ App/Patternaut/PatternautApp.swift
@@ -3,9 +3,13 @@ import SwiftUI
 @main
 struct PatternautApp: App {
     var body: some Scene {
-        WindowGroup("Patternaut") {
-            ContentView()
+        DocumentGroup(newDocument: { PatternautDocument() }) { file in
+            ContentView(model: file.document.model)
         }
-        .windowStyle(.titleBar)
+        .commands {
+            CommandGroup(replacing: .help) {
+                Link("Patternaut Website", destination: URL(string: "https://patternaut.iamjarl.com")!)
+            }
+        }
     }
 }